+++ tests/testthat/test_fixed_effects_in_E_step.R
@@ -88,6 +88,7 @@ test_that("Works with second order random walk and logit model", {
   fit <- fit[c("state_vars", "state_vecs", "Q", "lag_one_cov")]
   # save_to_test(fit, "E_step_sim_EKF_order_two")
 
+
   expect_equal(fit, read_to_test("E_step_sim_EKF_order_two"), tolerance = 0.1)
 })
 
@@ -101,7 +102,7 @@ test_that("Get loglike to work so you can use verbose with E-step fixed effects"
                                    save_risk_set = F, save_data = F,
                                    LR = 1 / 1.25),
                     verbose = 5)
-    , regexp = "Iteration\\s+\\d+\\sended with conv criteria\\s+\\d+.\\d+\\s+The log likelihood is\\s+")
+    , regexp = "Iteration\\s+\\d+\\sended with conv criteria\\s+\\d+.\\d+\\s+The log likelihood of the mean path is\\s+")
 })