+++ tests/testthat/test_residuals.R
@@ -137,47 +137,51 @@ test_that("State space error gives correct dimension with fixed effects", {
 # Data sets used in vignettes
 
 # PBC dataset described in Fleming & Harrington (1991)
-fit <- ddhazard(
-  formula = survival::Surv(time, status == 2) ~ log(protime),
-  data = pbc, Q_0 = diag(rep(1e8, 2)), by = 100,
-  Q = diag(rep(1e-2, 2)), max_T = 3600,
-  a = rep(0, 2),
-  control = list(est_Q_0 = F, eps = .1))
+local({
+  pbc <- pbc[!is.na(pbc$protime), ]
 
-test_that("Pearson residuals and raw residuals for logistic model are consistent with each other", {
-  pearson_res <- residuals(object = fit, type = "pearson", data = pbc)
-  raw_res <- residuals(object = fit, type = "raw", data = pbc)
+  fit <- ddhazard(
+    formula = survival::Surv(time, status == 2) ~ log(protime),
+    data = pbc, Q_0 = diag(rep(1e8, 2)), by = 100, id = pbc$id,
+    Q = diag(rep(1e-2, 2)), max_T = 3600,
+    a = rep(0, 2),
+    control = list(est_Q_0 = F, eps = .1))
 
-  expect_s3_class(pearson_res, "fahrmeier_94_res")
-  expect_s3_class(raw_res, "fahrmeier_94_res")
+  test_that("Pearson residuals and raw residuals for logistic model are consistent with each other", {
+    pearson_res <- residuals(object = fit, type = "pearson", data = pbc)
+    raw_res <- residuals(object = fit, type = "raw", data = pbc)
 
-  expect_equal(pearson_res$type, "pearson")
-  expect_equal(raw_res$type, "raw")
+    expect_s3_class(pearson_res, "fahrmeier_94_res")
+    expect_s3_class(raw_res, "fahrmeier_94_res")
 
-  for(i in seq_along(pearson_res$residuals))
-    expect_equal(pearson_res$residuals[[i]][, "residuals"],
-                 raw_res$residuals[[i]][, "residuals"]/
-                   sqrt(raw_res$residuals[[i]][, "p_est"] * (1 - raw_res$residuals[[i]][, "p_est"])))
-})
+    expect_equal(pearson_res$type, "pearson")
+    expect_equal(raw_res$type, "raw")
 
-test_that("Cases in residuals match cases in data", {
-  is_case <- pbc$status == 2 & pbc$time <= 3600
-  pearson_res <- residuals(object = fit, type = "pearson", data = pbc)
+    for(i in seq_along(pearson_res$residuals))
+      expect_equal(pearson_res$residuals[[i]][, "residuals"],
+                   raw_res$residuals[[i]][, "residuals"]/
+                     sqrt(raw_res$residuals[[i]][, "p_est"] * (1 - raw_res$residuals[[i]][, "p_est"])))
+  })
 
-  is_case_residuals <- rep(0, nrow(pbc))
-  for(i in seq_along(pearson_res$residuals)){
-    y_n_index <- data.frame(pearson_res$residuals[[i]][, c("residuals", "Y", "row_num")])
+  test_that("Cases in residuals match cases in data", {
+    is_case <- pbc$status == 2 & pbc$time <= 3600
+    pearson_res <- residuals(object = fit, type = "pearson", data = pbc)
 
-    is_case_residuals[y_n_index$row_num[y_n_index$Y == 1]] =
-      is_case_residuals[y_n_index$row_num[y_n_index$Y == 1]] + 1
+    is_case_residuals <- rep(0, nrow(pbc))
+    for(i in seq_along(pearson_res$residuals)){
+      y_n_index <- data.frame(pearson_res$residuals[[i]][, c("residuals", "Y", "row_num")])
 
-    expect_equal(sum(y_n_index$residuals[y_n_index$Y == 1] > 0),
-                 sum(y_n_index$Y == 1))
-    expect_equal(sum(y_n_index$residuals[y_n_index$Y == 0] < 0),
-                 sum(y_n_index$Y == 0))
-  }
+      is_case_residuals[y_n_index$row_num[y_n_index$Y == 1]] =
+        is_case_residuals[y_n_index$row_num[y_n_index$Y == 1]] + 1
 
-  expect_equal(is_case_residuals, is_case + 0)
+      expect_equal(sum(y_n_index$residuals[y_n_index$Y == 1] > 0),
+                   sum(y_n_index$Y == 1))
+      expect_equal(sum(y_n_index$residuals[y_n_index$Y == 0] < 0),
+                   sum(y_n_index$Y == 0))
+    }
+
+    expect_equal(is_case_residuals, is_case + 0)
+  })
 })